pyproject.toml
[project]
name = "n8n-workflow-system"
version = "1.0.0"
description = "n8n Workflow Popularity System"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "asyncpg==0.29.0",
    "pydantic==2.5.0",
    "httpx==0.25.2",
    "google-api-python-client==2.108.0",
    "pytrends==4.9.2",
    "apscheduler==3.10.4",
    "python-multipart==0.0.6",
    "python-levenshtein==0.23.0",
    "python-dotenv==1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1"
]